Create ButterflyPattern.java
This program first takes the number of rows for the butterfly pattern as input from the user using the Scanner class. It then uses nested loops to generate the upper and lower parts of the butterfly pattern. The number of asterisks and spaces is adjusted based on the row number to create the desired pattern.

Compile and run this program, and you will get a butterfly pattern based on the number of rows you input.